Use long int in cpp quantize only shift calculation

+++ ModelOptimizations/DlQuantization/src/trim_functions.cu
@@ -59,7 +59,9 @@ template <typename DTYPE>
 __global__ void quantizeToFxpKernel(const DTYPE* in, int cnt, const TfEncoding encoding, DTYPE* out,
                                     RoundingMode rounding_mode, bool shiftToSigned)
 {
-    int shift = 0;
+    // Using long int to account for case of signed symmetric 32 bit, when shift will be 2^31 which regular 8 bit int
+    // cannot handle.
+    long int shift = 0;
     if (shiftToSigned) {
         shift = pow(2, encoding.bw - 1);
     }
